# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  upcoming: "upcoming",
  live: "live",
  ended: "ended",
} as const;

export interface Project {
  id: string;
  name: string;
  ticker: string;
  description: string;
  logoUrl?: string;
  bannerUrl?: string;
  tokenAddress?: string;
  status: ProjectStatus;
  totalRaise: number;
  raisedAmount: number;
  tokenPrice: number;
  minAllocation?: number;
  maxAllocation?: number;
  startDate: string;
  endDate: string;
  websiteUrl?: string;
  twitterUrl?: string;
  telegramUrl?: string;
  participants: number;
  category: string;
  createdAt: string;
}

export interface CreateProjectBody {
  name: string;
  ticker: string;
  description: string;
  tokenAddress?: string;
  totalRaise: number;
  tokenPrice: number;
  minAllocation?: number;
  maxAllocation?: number;
  startDate: string;
  endDate: string;
  websiteUrl?: string;
  twitterUrl?: string;
  telegramUrl?: string;
  category: string;
}

export interface StakingTier {
  id: string;
  name: string;
  requiredAmount: number;
  allocationMultiplier: number;
  benefits: string[];
  color: string;
}

export interface StakingPosition {
  id: string;
  walletAddress: string;
  amountStaked: number;
  tier: string;
  rewardsEarned: number;
  stakedAt: string;
}

export interface CreateStakingPositionBody {
  walletAddress: string;
  amountStaked: number;
}

export interface PlatformStats {
  totalProjectsLaunched: number;
  totalRaisedUsd: number;
  totalStakers: number;
  totalDpinoStaked: number;
  activeLaunches: number;
  successRate: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListProjectsParams = {
  status?: ListProjectsStatus;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  upcoming: "upcoming",
  live: "live",
  ended: "ended",
  all: "all",
} as const;

export type ListStakingPositionsParams = {
  walletAddress?: string;
};