# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all projects on the launchpad
 * @summary List all launchpad projects
 */
export const ListProjectsQueryParams = zod.object({
  status: zod.enum(["upcoming", "live", "ended", "all"]).optional(),
});

export const ListProjectsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  ticker: zod.string(),
  description: zod.string(),
  logoUrl: zod.string().optional(),
  bannerUrl: zod.string().optional(),
  tokenAddress: zod.string().optional(),
  status: zod.enum(["upcoming", "live", "ended"]),
  totalRaise: zod.number(),
  raisedAmount: zod.number(),
  tokenPrice: zod.number(),
  minAllocation: zod.number().optional(),
  maxAllocation: zod.number().optional(),
  startDate: zod.coerce.date(),
  endDate: zod.coerce.date(),
  websiteUrl: zod.string().optional(),
  twitterUrl: zod.string().optional(),
  telegramUrl: zod.string().optional(),
  participants: zod.number(),
  category: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Submit a project for listing
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  ticker: zod.string(),
  description: zod.string(),
  tokenAddress: zod.string().optional(),
  totalRaise: zod.number(),
  tokenPrice: zod.number(),
  minAllocation: zod.number().optional(),
  maxAllocation: zod.number().optional(),
  startDate: zod.coerce.date(),
  endDate: zod.coerce.date(),
  websiteUrl: zod.string().optional(),
  twitterUrl: zod.string().optional(),
  telegramUrl: zod.string().optional(),
  category: zod.string(),
});

/**
 * @summary Get a single project
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.string(),
});

export const GetProjectResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ticker: zod.string(),
  description: zod.string(),
  logoUrl: zod.string().optional(),
  bannerUrl: zod.string().optional(),
  tokenAddress: zod.string().optional(),
  status: zod.enum(["upcoming", "live", "ended"]),
  totalRaise: zod.number(),
  raisedAmount: zod.number(),
  tokenPrice: zod.number(),
  minAllocation: zod.number().optional(),
  maxAllocation: zod.number().optional(),
  startDate: zod.coerce.date(),
  endDate: zod.coerce.date(),
  websiteUrl: zod.string().optional(),
  twitterUrl: zod.string().optional(),
  telegramUrl: zod.string().optional(),
  participants: zod.number(),
  category: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get all staking tiers
 */
export const GetStakingTiersResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  requiredAmount: zod.number(),
  allocationMultiplier: zod.number(),
  benefits: zod.array(zod.string()),
  color: zod.string(),
});
export const GetStakingTiersResponse = zod.array(GetStakingTiersResponseItem);

/**
 * @summary Get staking positions
 */
export const ListStakingPositionsQueryParams = zod.object({
  walletAddress: zod.coerce.string().optional(),
});

export const ListStakingPositionsResponseItem = zod.object({
  id: zod.string(),
  walletAddress: zod.string(),
  amountStaked: zod.number(),
  tier: zod.string(),
  rewardsEarned: zod.number(),
  stakedAt: zod.coerce.date(),
});
export const ListStakingPositionsResponse = zod.array(
  ListStakingPositionsResponseItem,
);

/**
 * @summary Stake $DPINO
 */
export const CreateStakingPositionBody = zod.object({
  walletAddress: zod.string(),
  amountStaked: zod.number(),
});

/**
 * @summary Get platform-wide statistics
 */
export const GetPlatformStatsResponse = zod.object({
  totalProjectsLaunched: zod.number(),
  totalRaisedUsd: zod.number(),
  totalStakers: zod.number(),
  totalDpinoStaked: zod.number(),
  activeLaunches: zod.number(),
  successRate: zod.number(),
});

/**
 * @summary Get featured/trending projects
 */
export const GetFeaturedProjectsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  ticker: zod.string(),
  description: zod.string(),
  logoUrl: zod.string().optional(),
  bannerUrl: zod.string().optional(),
  tokenAddress: zod.string().optional(),
  status: zod.enum(["upcoming", "live", "ended"]),
  totalRaise: zod.number(),
  raisedAmount: zod.number(),
  tokenPrice: zod.number(),
  minAllocation: zod.number().optional(),
  maxAllocation: zod.number().optional(),
  startDate: zod.coerce.date(),
  endDate: zod.coerce.date(),
  websiteUrl: zod.string().optional(),
  twitterUrl: zod.string().optional(),
  telegramUrl: zod.string().optional(),
  participants: zod.number(),
  category: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetFeaturedProjectsResponse = zod.array(
  GetFeaturedProjectsResponseItem,
);