# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectBody {
  name: string;
  ticker: string;
  description: string;
  tokenAddress?: string;
  totalRaise: number;
  tokenPrice: number;
  minAllocation?: number;
  maxAllocation?: number;
  startDate: Date;
  endDate: Date;
  websiteUrl?: string;
  twitterUrl?: string;
  telegramUrl?: string;
  category: string;
}