# lib/api-zod/src/generated/types/createStakingPositionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStakingPositionBody {
  walletAddress: string;
  amountStaked: number;
}