# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalProjectsLaunched: number;
  totalRaisedUsd: number;
  totalStakers: number;
  totalDpinoStaked: number;
  activeLaunches: number;
  successRate: number;
}