# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from "./projectStatus";

export interface Project {
  id: string;
  name: string;
  ticker: string;
  description: string;
  logoUrl?: string;
  bannerUrl?: string;
  tokenAddress?: string;
  status: ProjectStatus;
  totalRaise: number;
  raisedAmount: number;
  tokenPrice: number;
  minAllocation?: number;
  maxAllocation?: number;
  startDate: Date;
  endDate: Date;
  websiteUrl?: string;
  twitterUrl?: string;
  telegramUrl?: string;
  participants: number;
  category: string;
  createdAt: Date;
}