# lib/api-zod/src/generated/types/stakingTier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DPINO Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StakingTier {
  id: string;
  name: string;
  requiredAmount: number;
  allocationMultiplier: number;
  benefits: string[];
  color: string;
}